# Trustify.Admin/src/app/api/models/group-dto.ts
/* tslint:disable */
export interface GroupDTO {
  clientRoles?: null | {[key: string]: Array<string>};
  id?: null | string;
  name?: null | string;
  path?: null | string;
  realmRoles?: null | Array<string>;
}