# Trustify.Admin/src/app/api/models/group-roles-wrapper.ts
/* tslint:disable */
import { RoleWrapper } from './role-wrapper';
export interface GroupRolesWrapper {
  clientId?: null | string;
  groupId?: null | string;
  roleWrappers?: null | Array<RoleWrapper>;
}